Fix various bugs concerning skipping and list navigation.
The whole concept of how list fragments are cached and how this cache
is filled has been revised.

In previous versions, a strict ordering of request, retrieval of data,
and handling of data had to be maintained. It was possible to do this
because of the old way of handling requests, i.e., having only one of
them around. It was not easy, but it worked, somewhat.

Now, there can be several active requests. Cancellation of requests
was a big problem in previous versions already, but with the new code
base, our whole house of cards has finally collapsed. This patch fixes
the cache structure by moving the data from the single piece of memory
in the D-Bus list to multiple places, the viewports. The complicated
concept of cache modification structures and dependence on ordering
has been dropped. Instead, we have list segment fetchers now which
fill viewports. These are easier to understand and to maintain. Since
they are not part of D-Bus lists, there can be multiple viewports for
a single list, further supporting concurrency.

While the code was rewritten to make use of viewports, it became
apparent that there were multiple potential users of the same list,
each requiring their own viewport:
- list navigation by user
- list crawler for playback and prefetch
- skip requests
- binary search

Change-Id: Iec176c034cdbe6ce43bb79608277a69c733aa185