filebrowser: Fix bad display content and crash on login error.
The software showed kind of a frozen display when logging into an
external music service failed (always "Accessing..."), and sometimes
it would even crash.

This patch fixes the display content (now "Unavailable" instead of
"Accessing..."), makes sure that updating the busy flag doesn't
overwrite the error message sent to the ARM too soon, and avoids the
crash (assertion because of bad jump-to-context state).

Change-Id: I2accd86ab0e4014f6a0e2cda3b63e7a172a3c41a