view mananger: Add update mode FULL_SERIALIZE_NEEDED.
Also change behavior of FORCE_SERIALIZE, which now always serializes
the full view, not just an update. This makes sense because this mode
is used when switching to a new view or sending a view by force,
potentially replacing a different view.

Mode UPDATE_NEEDED still behaves as it did before, i.e., it sends an
update of those portions of the view which require an update.

Mode FULL_SERIALIZE_NEEDED sends a full view if the view is active.

Change-Id: Id3dbbe6fa2083f28b9b43938a4c3f6133d844490