player: Refine delay of prefetch operations (#1504).
We do not always have to delay prefetching. In pause mode, there is no
stuttering, so we can prefetch as quickly as possibly.

Note that this is not a full fix for #1504. The delay timer and its
effect of skipping in case of playback failure described in the ticket
is still slowing down skips over stretches of unplayable streams.

See also d913692b185187c80395f90a90942fce18e30f84 (#1426).

Change-Id: I3a6788dc93c245c375cb714036b6a642b8e61b4f