filebrowser: Fix two crashes (#1469).
The assertion in view_filebrowser.cc has been haunting us for a long
time. I was lucky to run into it with a debug version of drcpd and
thus to obtain a usable coredump to track it down.

While trying to reproduce the problem, I also ran into the now-removed
assertion in view_filebrowser_airable.cc.

Change-Id: I98e640130ad8693c682477568cae392d14d8cb50