filebrowser: Giving up on partial display updates.
The SPI slave won't accept partial updates, so clean up a bit and
provide a place for documenting the problem.

Also fix serialization in case of error by not leaving the XML writer
early and allowing it to put the last few values after the lines.

Don't query live busy state for the logs. It is irrelevant now, so we
don't need to dump it. The Busy::is_busy() function is only used in
unit tests now.

Change-Id: Iff6ce566a23bb306c43986d0e58a6f5f064ad017